+++ packages/mui/src/CheckboxesWidget/CheckboxesWidget.tsx
@@ -10,16 +10,16 @@ import {
   StrictRJSFSchema,
 } from "@rjsf/utils";
 
-const selectValue = (value: any, selected: any, all: any) => {
+const selectValue = (value: any, selected: any[], all: any) => {
   const at = all.indexOf(value);
   const updated = selected.slice(0, at).concat(value, selected.slice(at));
 
   // As inserting values at predefined index positions doesn't work with empty
   // arrays, we need to reorder the updated selection to match the initial order
-  return updated.sort((a: any, b: any) => all.indexOf(a) > all.indexOf(b));
+  return updated.sort((a, b) => Number(all.indexOf(a) > all.indexOf(b)));
 };
 
-const deselectValue = (value: any, selected: any) => {
+const deselectValue = (value: any, selected: any[]) => {
   return selected.filter((v: any) => v !== value);
 };
 
@@ -47,6 +47,7 @@ export default function CheckboxesWidget<
   onFocus,
 }: WidgetProps<T, S, F>) {
   const { enumOptions, enumDisabled, inline } = options;
+  const checkboxesValues = Array.isArray(value) ? value : [value];
 
   const _onChange =
     (option: any) =>
@@ -54,9 +55,9 @@ export default function CheckboxesWidget<
       const all = (enumOptions as any).map(({ value }: any) => value);
 
       if (checked) {
-        onChange(selectValue(option.value, value, all));
+        onChange(selectValue(option.value, checkboxesValues, all));
       } else {
-        onChange(deselectValue(option.value, value));
+        onChange(deselectValue(option.value, checkboxesValues));
       }
     };
 
@@ -75,7 +76,7 @@ export default function CheckboxesWidget<
       <FormGroup id={id} row={!!inline}>
         {Array.isArray(enumOptions) &&
           enumOptions.map((option, index: number) => {
-            const checked = value.indexOf(option.value) !== -1;
+            const checked = checkboxesValues.includes(option.value);
             const itemDisabled =
               Array.isArray(enumDisabled) &&
               enumDisabled.indexOf(option.value) !== -1;